a better function api
previously user scripts needed to take `config` as a first argument, and they were
passed any relevant region and group specific config (for example the list of kafka brokers).
now we also want to pass the `region` in. this revealed a problem with the previous api,
as it was not easy to modify in a backward-compatible way. now functions take no config
argument. instead they have access to a context object that contains just the region and the
group config. this can also be extended in the future if we want to add properties here
in a backward compatible way which requires no changes to existing functions.