improve chart (#69)
the chart example previously was pretty janky, it just looked for a
field called "date" and used that as the x-axis. now the user can choose
which field they want to use as the axis.

also adds an example script and moves it to it's own component. this
structure should make it easier to add more visualization types in
future.