custom function parameters and input types (#78)
this adds the ability for users to customize the
function input types in the UI, as well as support different python types (not just strings) as arguments. now all inputs to script have to a subclass of FunctionParameter and implement its methods

this enables us to define a lot more custom behavior around the inputs. for example we can:
- have numeric types like float and int, with validation on the UI as well
- have custom input types for the same python type. e.g. now there is a regular text input as well as multi-line textarea for longer inputs.
- we can add other stuff in the future such as multiple select and custom autocompletes