test utilities for autocomplete
now the `execute_with_context` test utility will check a function to see if
it has any dynamically autocompleted parameters, and generate the
autocompletions if so.

it asserts that all autocomplete functions return a list of strings,
and returns the values to the user so that can make test specific
assertions on them.