fix: Fix circular dependency (#94)
previously the decorators and blueprint files were importing app.py, and
app.py was also importing the decorators and blueprint files. delayed
imports was being used to get around it, but this is a workaround that
really shouldn't be needed.

since app.py is the top level of the flask application, no one else
should be importing it.