split up the script result component (#99)
this is the first step to implementing a standard interface for react
components that display results of script executions.

this will enable them to be customizable in the future, so users can
contribute their own custom visualizations.

there are no functional changes or improvements here (that will come
later) - this pr just shifts code mostly as-is from a single big
ScriptResult component with lots of nested conditions to dedicated
components for each visualization which all get passed the exact same
props.