add approval policy and storage abstract classes
the approval policy allows the user of this library to provide
custom functionality to check whether approval is required for a function.

custom approval functionality is still a work in progress. while it is
being built out we can set the policy to `allow_all` in order to mimic
existing behavior.