add approval policy and storage abstract classes (#101)
this is an initial step to enabling write operations in the script
runner. it introduces some abstract classes and (not yet implemented)
api endpoints.

there are two new abstract classes
- an approval policy: allows the user of this library to provide custom
functionality to check whether approval is required for a function.
- an approval store: the user can plug any storage layer they want into
it

this also removes the global `config` state so that the flask app can be
more easily tested. it enables multiple script-runner apps with
different config to be run together, as part of a single python
application as there are no more globals in use

custom approval functionality is still a work in progress. while it is
being built out we can set the policy to `allow_all` in order to
preserve existing behavior.